Refactor chirpstack events
